requirements.txt
paddlepaddle
paddleocr 
fastapi
uvicorn
ultralytics
opencv-python
python-multipart
pytest
pytest-cov
google-generativeai